requirements.txt
# Dashboard Manutenção - Dependências

# Core
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.18.0

# BigQuery
google-cloud-bigquery>=3.13.0
db-dtypes>=1.3.0
pyarrow>=14.0.0

# Google Sheets
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Agendamento
schedule>=1.2.0

# HTTP requests
requests>=2.31.0

# CLI
click>=8.1.0
